pyproject.toml
[project]
name = "ai-stock-agent"
version = "0.1.0"
description = "Autonomous AI Stock Research Agent with multi-tool integration"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "langgraph>=1.0.0",
    "langchain-core>=1.0.0",
    "openai>=1.3.0",
    "supabase>=2.0.0",
    "psycopg[binary]>=3.1.0",
    "pgvector>=0.2.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "structlog>=23.2.0",
    "jinja2>=3.1.2",
    "markdown>=3.5.1",
    "weasyprint>=60.1",
    "apscheduler>=3.10.4",
    "feedparser>=6.0.10",
    "scikit-learn>=1.3.2",
    "numpy>=1.26.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-httpx>=0.27.0",
    "freezegun>=1.2.2",
]

[project.optional-dependencies]
dev = [
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "memory", "apps.api"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

